Move to float until FFT
Move to float until FFT.
Still, FFT outputs double in this commit.

+++ src/fft.h
@@ -3,7 +3,7 @@
 
 #include "fftw3.h"
 
-void fft (double *data, uint64_t dataSamples, double complex *output);
+void fft (float *data, uint64_t dataSamples, double complex *output);
 void ifft (double complex *data, uint64_t dataSamples, double *output);
 void fftToFileHalf (double *data, uint64_t dataSamples, double sampleRate, FILE *fptr);
 void fftToFile (double *data, uint64_t dataSamples, double sampleRate, FILE *fptr);
